WIP: adding filters to course list * Refactored filter into checkbox and drop down filters
