+++ analytics_dashboard/static/apps/components/filter/templates/checkbox-filter.underscore
@@ -1,0 +1,12 @@
+<% if (!_.isEmpty(filterValues)) { %>
+    <% var filterId = 'filter-' + filterKey; %>
+    <hr>
+    <div id="<%- filterId %>">
+        <% _.each(filterValues, function (filterValue) { %>
+            <% if (filterValue.name === 'inactive') { %>
+                <input id="<%- filterValue.name %>" type="checkbox" value="<%- filterValue.name %>" <% if (isChecked) { %> checked <% } %>>
+                <label for="<%- filterValue.name %>"> <%- selectDisplayName %> </label><br>
+            <% } %>
+        <% }); %>
+    </div>
+<% } %>
